+++ Library/Services/INfieldSurveyInterviewerAssignmentsService.cs
@@ -13,6 +13,7 @@
 //    You should have received a copy of the GNU Lesser General Public License
 //    along with Nfield.SDK.  If not, see <http://www.gnu.org/licenses/>.
 
+using Nfield.Models;
 using System.Threading.Tasks;
 
 namespace Nfield.Services
@@ -43,5 +44,17 @@ namespace Nfield.Services
         /// <exception cref="Nfield.Exceptions.NfieldErrorException"></exception>
         /// <exception cref="Nfield.Exceptions.NfieldHttpResponseException"></exception>
         Task UnassignAsync(string surveyId, string interviewerId);
+
+        /// <summary>
+        /// Assign an interviewer to a Survey or sampling points in a survey
+        /// </summary>
+        /// <param name="surveyId">The id of the survey from which to unassign the interviewer.</param>         
+        /// <param name="interviewerId">The interviewer id.</param>
+        /// <param name="model">Assignment sata</param>
+        /// <exception cref="T:System.AggregateException"></exception>
+        /// The aggregate exception can contain:
+        /// <exception cref="Nfield.Exceptions.NfieldErrorException"></exception>
+        /// <exception cref="Nfield.Exceptions.NfieldHttpResponseException"></exception>
+        Task PutAsync(string surveyId, string interviewerId, SurveyInterviewerAssignmentModel model);
     }
 }
